Icecube: config: Bumped bsp version to v3.4.0-1 (#583)
Summary:
# Description

This PR is for the icecube platform service to support the latest BSP version 3.4.0-1.

# Motivation

According to the latest BSP tag, we need to adjust the BSP version to 3.4.0-1 in the platform config file.

Pull Request resolved: https://github.com/facebook/fboss/pull/583

Test Plan:
The format has been verified as correct on the JSONLint website.
Used the jq command to pretty-print the format.
Test log as follows:

<img width="2480" height="878" alt="image" src="https://github.com/user-attachments/assets/314371c8-b784-4c85-9c39-434db7a3a4f4" />

[icecube_pm_test_log_10_9.txt](https://github.com/user-attachments/files/22840258/icecube_pm_test_log_10_9.txt)

Reviewed By: ezeob002

Differential Revision: D84721147

Pulled By: mikechoifb

fbshipit-source-id: a82fc777f5121e8203fd7544fae02bfb8706fec3