Kill MutexWrapper
Summary: `fb303` uses `SharedMutex` for its `Synchronized` objects, but it does so through the `MutexWrapper` adapter, so that all the `lock()` calls did not have to be codemodded to `wlock()`. However, in some codepaths it is actually possible to use an rlock, so it's best to remove the wrapper and just codemod all the call sites.

Reviewed By: uvdn7

Differential Revision: D84917811

fbshipit-source-id: 33c3fbee5d9058c027da821bca30f2e725c124a5