Laser status addition onto thrift file.
Summary:
This diff adds a new field to the transceiver.thrift file, which is used to represent the tunable laser status. The new field is called TunableLaserStatus and includes three sub-fields:
TuningStatus: The TuningStatus field indicates whether the laser is currently being tuned or not
WavelengthLockingStatus: WavelengthLockingStatus field indicates whether the laser is currently locked to a specific wavelength or not.
Laser frequency : The current operational laser frequency
The diff also fixes the enum ordering for the frequency grid, now its aligning with CMIS ordering

Reviewed By: xanabani

Differential Revision: D83987931

fbshipit-source-id: 0068c281c356ad27cf0a9204d4aa788828b28908