Clear nextDsfUpdate when enqueuing GR expiry event
Summary:
The following sequence of update event will lead a race and hence clearing all remote neighbors on a device.

1. DsfUpdate of remote interface and neighbors update from a remote device
2. GR timer expiry of the remote device, leading to the enqueue of neighbor clearing
3. While both event 1 and 2 are not yet processed, remote device came up and send a new DsfUpdate of remote interfaces.

Due to the optimization of next dsf update, we will not enqueue another event if there's already a DsfUpdate (1) in the queue. 3 will then modify the content of the update of 1, but in the event base queue, it has

Head -> [DsfUpdate(enqueued by 1 but has content from 3), GR expiry(enqueued by 2)].

The net effect is all neighbors will get cleared out because of this. To fix the issue, we need to clear `nextDsfUpdate_` while enqueuing GR expiry. This will force following updates to queue another event to the event base.

Reviewed By: shri-khare

Differential Revision:
D85073824

Privacy Context Container: L1297311

fbshipit-source-id: da6ef9cf9a5b8ec6990b59d0d1d0e48c70c8454f