UT to repro race condition in DsfSubscription observed in S575581
Summary:
The following sequence of update event will lead a race and hence clearing all remote neighbors on a device.

1. DsfUpdate of remote interface and neighbors update from a remote device
2. GR timer expiry of the remote device, leading to the enqueue of neighbor clearing
3. While both event 1 and 2 are not yet processed, remote device came up and send a new DsfUpdate of remote interfaces.

Due to the optimization of next dsf update, we will not enqueue another event if there's already a DsfUpdate (1) in the queue. 3 will then modify the content of the update of 1, but in the event base queue, it has

Head -> [DsfUpdate(enqueued by 1 but has content from 3), GR expiry(enqueued by 2)].

The net effect is all neighbors will get cleared out because of this. See next diff to fix the issue, but here we are creating a UT to repro the same sequence.

Reviewed By: shri-khare, nivinl

Differential Revision:
D85105882

Privacy Context Container: L1297311

fbshipit-source-id: 467f04c1acec7d832b454c839d05b15fc148ccbc