Towards a alternative syncFib approach
Summary:
Current syncFib for a clientID (e.g. bgp) approach
- Iterate over the routes, clear nhops for this client ID. If
this is the only clientId, mark this route for deletion.
- Now add all routes that came in with syncFib.

The problem with this approach is that this gets rid off all the override
(nhops or backup ecmp switching mode info from the routes) as we delete
the routes. This will happen even if the route has had no change in
nhops.

This approach is different from what we do with add/update routes
approach. In the latter we just update the route if the route has
changed for this client ID.

Over the next few diffs we will unify the 2 approaches. By doing the
following on syncFib
i. Add/update routes as is done for add/update API
ii. After this is done, clear out routes that were not touched as
part of add/update

Reviewed By: zechengh09

Differential Revision:
D85122105

Privacy Context Container: L1297311

fbshipit-source-id: 18ec265724f45f90a809314d2ca90c31d01a7fcc