Stop using state modifier abstraction for ERM and Shel Manager
Summary:
**Background**
There is a fundamental issue with state modifier. Imagine we have state modifiers M1, M2 and M3. The initial state delta prior to state modifier is StateDelta<oldState, newState>. For the sake of simplicity, let's say all three modifiers don't create more than 1 stateDelta.

M1 will consume the state delta and let's say modify the state delta to be <oldState, newState1>
M2 will consume the above state delta and modify it to <oldState, newState2>
M3 will consume the above state delta and modify it to <oldState, newState3>

Now if modifier is keeping track of counters or some other state from StateDelta, M1 will not be able to see the state change from newState2 and newState3. This is the same for M2 not being able to see the delta in newState3.

Therefore, scrubbing off this abstraction and don't allow more modifiers to be added. Shel manager will maintain as it's required for efficient ECMP shrink. ERM will be moved to FIB while computing routes.

**This diff**
Stop using the abstraction of state modifier, but directly invoke methods of ERM and shel manager to process the updates.

Reviewed By: jasmeetbagga

Differential Revision:
D85178729

Privacy Context Container: L1297311

fbshipit-source-id: fd99559d470192f83e5dcf98e0a2914e24f8728e