Override nhops should only be seen during rollback or reconstruction
Summary:
Override nhops should only be seen during rollback or reconstruction.
If we see it otherwise - i.e. with a route update not being done with
a rollback, this would be a bug. So
- DCHECK in tests
- Error log and ignore override in prod.
However we don't expect this to happen, this is mostly a defensive
play.

Reviewed By: zechengh09

Differential Revision:
D85211366

Privacy Context Container: L1297311

fbshipit-source-id: 341d7fab03a9f7a71b46854ffb0c51b9b9f773b3