Throw rather than abort if we cant find any candidateMergeGroups
Summary:
Rather than aborting, throw an exception back. While with the
current config (100% ecmp compression threshold), we should
always be able to find groups to merge. However, when we start
putting in other thresholds, its possible we may not be
able to find a group which meets this threshold.

Reviewed By: zechengh09

Differential Revision:
D85211892

Privacy Context Container: L1297311

fbshipit-source-id: 62cd8eaa407f69e85dad2330dcf07e439eac0a48