FwIsolate with numActiveLinks = 0: avoid leaving devices DRAINED
Summary:
See detailed comments in the code.
Especially Note for case (B).

Prior to this, the code compared number of Active fabric links at FwIsolate
with thresholds and could incorrectly leave the system DRAINED. This is illustrated in P2008718284.

Fixed here as detailed in the comment.

Reviewed By: jasmeetbagga

Differential Revision:
D85530206

Privacy Context Container: L1297311

fbshipit-source-id: 014236614416cd97172d57736ab59d74cc4fcd5b