Updating hw stats queue from unbounded to bounded queue of size 1000.
Summary:
In split agent architecture, hardware stats collected by a hw_agent, is pushed to sw-agent using a unbounded queue which could cause unwated high memory usage.  Following change addresses it by changing it to a bounded queue of size '1000'.

It's still defined in SPSC mode, perhaps it needs to be moved to MPSC mode in future.

Reviewed By: msomasundaran

Differential Revision: D84943501

fbshipit-source-id: 2358c0d739dc51547141d4df1b3f763d4f0db7e1