Fix coroutine object getting killed before handler resume in OSS (with GCC compiler)
Summary:
When we invoke co_await in a function foo() for ex, coroutines captures the state of the function and creates an awaitable object (https://www.internalfb.com/metamate/M290UD), stores the local variables and then returns the context back to the caller of foo. This awaitable object will be used to resume the coroutine at a later point (when the corresponding co_return is called). If for some reason, the stored variables go out of scope, then they are destroyed before the coroutine is resumed and this results in use after free error(some bad examples can be found at https://en.cppreference.com/w/cpp/language/coroutines.html)

if a function  calls co_await with ternary operator,
```
co_await cond() ? bar() : baz();
```

or if co_await is called with

```
co_await foo(bar())
```

then the lifetime of the resulting variable is temporary and goes out of scope once the expression ends(and the function is awaited). When the co_await is resumed, then we run into use after free error causing the crash observed in OSS (P1663576357).

In cases like these, we can extend the lifetime of these variables by explicitly materializing them with a temporary variable.

For ex:

```
auto temp = cond() ? bar() : baz(); <- where the result is captured in a temporary variable and scope is extended

co_await temp;
```
(or)
```
auto baz = bar();
co_await foo(baz);
```

Clang still extended the lifetime of these variables and so we did not observe any crash internally. GCC however did not match the same behavior which is why it was consistently crashing in OSS but not internally

Reviewed By: joseph5wu

Differential Revision: D84937439

fbshipit-source-id: 1be76cc99a8e6c045c01effe10f63e6f6d57dcd6