Add function to compare captured and expected sFlow packets
Summary: This diff adds a function to compare captured and expected sFlow packets in the AgentSflowMirrorTest.cpp file. The function then parses the captured packet into a SflowPacketParsed struct and creates an expected SflowPacketParsed from the inputs. The function then compares the two structs to ensure that they match. This aligns with the comparison and masking done in AgentSflowMirrorTest.

Differential Revision: D83598622

fbshipit-source-id: 659645543630c5a50b96a87aa07bf70989ebff1e