Change log format for platformPortConfigOverrideMatcher
Summary:
tldr Our benchmark framework parses stdout of each test to find the benchmark values that determine if a test is successful or not; these values are written in JSON format. Log was added recently in D85119246 to print out a Thrift struct serialized as a JSON, but the benchmark doesn't know the difference between benchmark JSON and any other log JSON, so it was trying to parse this new log, which was causing some issues.

This diff just replaces curly brackets with parentheses to avoid parsing the new log.

Reviewed By: Protick666

Differential Revision: D85771091

fbshipit-source-id: 4961fe6a6e8ee98856ac5764b786315eeb566a37