Create static buffers for async logging
Summary:
**Background**

While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.
The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.
Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue.

To re-use the same async logger we built for BCM Cinter/SAI Replayer, refactor async logger such that it can also accommodate state delta logger.

**This Diff**
Similar to AsyncLogger, create static buffers such that, in the case of catastrophic failures we can still flush it in the terminate handler.

As the lifetime of static variable is longer than member variables, define those as static and will be used by AsyncLoggerbase.

Differential Revision: D85684990

fbshipit-source-id: ccc756745f9a985553797e46e95bc9bf38eb8344