Move state delta flags to agent features
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

Move it to AgentFeatures such that other files, (e.g. SwAgent) can access them.

Reviewed By: yebolai

Differential Revision: D85684996

fbshipit-source-id: 9a4655d0ae81db7c781f7847ca89bb2a46fae0c8