Start building StateDeltaLogger
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

Start building state delta logger in BUCK and open source cmake.

Reviewed By: yebolai

Differential Revision: D84961230

fbshipit-source-id: f98c1a3218d1330380b1ddb433cb66a7fdb146d8