Start updating ECMP overrides in rib processing itself
Summary:
Currently we go through a complicated dance of updating
ECMP overrides post facto from the updateEventBase. This
queues a async lambda on the RIB thread to update the
overrides made. This is somewhat complicated, and
hard to reason about.

This diff is not super efficient since it computes
delta over emptyState, latestState. Will make it more efficient
in stacked diffs

Reviewed By: ashutoshgrewal

Differential Revision:
D85309389

Privacy Context Container: L1297311

fbshipit-source-id: 2bda937e52a5730562e74cf7784dcc0e29930270