Delta logging and reconstruction test
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

As titled. Create tests to log in all three protocols, and then verify the reconstructed oper delta is the same as the initial oper delta.

Reviewed By: yebolai

Differential Revision: D85699972

fbshipit-source-id: 08bd189735ff561020ce82a73276867f25901f9e