Log state detla before and after state modifier
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

As titled. This would help us to understand how the state delta looks like before state modifier and after state modifier.

Reviewed By: yebolai

Differential Revision: D84963311

fbshipit-source-id: 38cd58a119342ab4c78894aba8d6a4b902ee3d26