Test remaining functionalities for stateDelta logger
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

Add remaining functional tests for stateDelta logger. This includes

1. Logging for multiple state Deltas
2. Logging for list of state Deltas,
3. Construction/Deconstruction when flag is not enabled.
4. Verify Protocol
5. Verify Generation number

Reviewed By: yebolai

Differential Revision: D85705881

fbshipit-source-id: ded5ce7ac21e63162f0183b48a8d3f42933e5fab