Base class of stateDeltaLogger UT
Summary:
**Background**
While debugging issues with ECMP Resource Manager stress testing, we noticed several crashes that are hard to reproduced when there is constant BGP flaps.

The inability to reproduce such issue makes it very hard to debug - we are only relying on our luck reproducing edge cases with ixia etc.

Rather, we can log all the state delta transitions happening on the box and trace every single delta transition. This would help us step-by-step to reproduce the issue. Note that this debugging tool should be test-use only. In production, we should consider using FSDB as the oper deltas are already streamed to FSDB (or we might decide to log within agent).

**This Diff**

Create base class for StateDeltaLoggerTest - set the correct flags and init temporary file for testing/verification.

Also create a helper to construct switch state with routes.

Reviewed By: yebolai

Differential Revision: D85699155

fbshipit-source-id: 2e029a5e832dfafd9889d21746ef6c4c686a322c