Fix async logger bug when log entry size larger than buffer size
Summary:
There's a bug in asyncLogger, where it's not able to handle log entry size that's larger than buffer size. This issue does not exist for BcmCinter/SaiReplayer, but more prominent as we're logging state delta (it could be big due to size of the state).

There are two options
1. Keep writing into buffer and force flush until all logs are in buffer
2. Directly write to file and skip writing to buffer.

2 is simpler approach and also easy to maintain. To achieve 1 we need some more complex locking logic to guarantee no other logs can be written into the buffer during this period of time.

We can revisit the proper buffer size for stateDetlaLogger, but need to fix the bug regardless.

Differential Revision:
D85966430

Privacy Context Container: L1297311

fbshipit-source-id: aed2ade982d6588cbd9cd54b6fc13b7394c3b020