Unit tests for log size larger than buffer size
Summary:
This is fixing the bug described in D85966430:

**The Bug**
There's a bug in asyncLogger, where it's not able to handle log entry size that's larger than buffer size. This issue does not exist for BcmCinter/SaiReplayer, but more prominent as we're logging state delta (it could be big due to size of the state).
There are two options
1. Keep writing into buffer and force flush until all logs are in buffer
2. Directly write to file and skip writing to buffer.

2 is simpler approach and also easy to maintain. To achieve 1 we need some more complex locking logic to guarantee no other logs can be written into the buffer during this period of time.
We can revisit the proper buffer size for stateDetlaLogger, but need to fix the bug regardless.

**This Diff**
Creates a UT for log size larger than buffer size - prior to the fix, test will hang till timeout since it will wait forever for available size that's larger than the buffer.

Differential Revision: D85966431

fbshipit-source-id: fa929c4cccad62f9927939c7d951d7f84a63e1a0