Add additional link count checks for L1 fabric devices
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
- Keep track of number of links per VD
- Add checks to ensure the number of links at the various layers in the various
deployments are as expected.

Reviewed By: shri-khare

Differential Revision: D83032814

fbshipit-source-id: 513d0d724095841fbfb36b713230a0117697642a