Fix the MAX switch ID being used to be 3903 instead of 3904
Summary:
Max switch ID is currently set to 3904. However, this is wrong as max switch ID should
be set to the next multiple of 32 and minus one from it. Given the current highest
switch ID we use in cluster is 3871, we should leave find the next multiple of 32 which is
3904 and minus one from it and use 3903 as the max switch ID so that SDK sets the
same in HW rather than trying to round again.

Reviewed By: shri-khare

Differential Revision: D83449533

fbshipit-source-id: 2162bc2ec4113e86f0046b5afa3cc5b56467bc7e