Implement packet TX on fabric port API for specific packet types
Summary:
Make common the hostifApi.send() API used for sync sending with a new parameter, packet type
added. A new attribute to specify packet type for the transmitted packets is introduced in the API.
This is for use when sending packets on fabric ports, so that various header fields could be set
based on the type of packet being sent like fabric link monitoring packet.

Reviewed By: shri-khare

Differential Revision: D84167134

fbshipit-source-id: f928e6fc486e0d4829450cfffd871c804c808416