Add an option to specify the packet type in packet tx thrift stream
Summary:
A new attribute to specify packet type for the transmitted packets. This is for use when sending
packets on fabric ports, so that various header fields could be set based on the type of packet
being sent like fabric link monitoring packet.

Add a new option in thrift stream to specify the packet type as well.

Reviewed By: shri-khare

Differential Revision: D84386001

fbshipit-source-id: 2b9997e790a91a67d342ddb210047899d22578bf