Introduce SaiPhyRetimer
Summary:
This diff introduces a new header file for SaiPhyRetimer

The class is aimed to be generic SAI-based retimer class designed to communicate with XPHY retimers. It will use the BspPhyIO instance we created for each phy io controller

Note that many methods are currently placeholders with TODOs. We will incrementally add support for them in the cpp file. Some we might not support soon (eg: getPortPrbs), but still have to define here as they are pure virtual functions in ExternalPhy  / HwSwitchCallback classes.

**Also ignore the unused lint warnings.**

Reviewed By: harshitgulati18

Differential Revision: D85173516

fbshipit-source-id: 0881e828286ec943966ac68bba5521319e201042