Define clearPortAsicPrbsStats in sai and bcm handler
Summary:
By implementing these changes, the Fboss Agent will be able to perform PRBS link tests in a multi-switch environment.

## This Diff

### Summary

This diff defines the `clearPortAsicPrbsStats` function in both SAI and BCM handlers, which is used to clear the ASIC PRBS statistics for a given port. The function is also implemented in the `HwSwitchThriftClientTable` to sync the `clearPortAsicPrbsStats` call to the corresponding switch.

Reviewed By: zechengh09

Differential Revision:
D85899658

Privacy Context Container: L1297311

fbshipit-source-id: c05e33eb0546a1f23da46f258e765a9752fa0267