Change weutil config file flag name
Summary:
Accton could not run OSS data_corral_service_hw_test because the weutil.json file is removed. This information is actually just moved to the platform_manager.json file so replacing that config with the PM config is sufficient to get the tests to run.

This diff just changes the config file flag name to make it make sense.

Reviewed By: RomanChoporov-Meta

Differential Revision: D86164716

fbshipit-source-id: fa182a7c8b00a60a4e8cac175cf165a375d38b9c