media interface code ZR_OIF_ZRA_800G full support
Summary: This diff adds full support for the media interface code ZR_OIF_ZRA_800G in the FBOSS QSFP service module. The changes include adding the new media interface code to the supported list in the CmisHelper.h and CmisHelper.cpp files, as well as updating the mapping between SMFMediaInterfaceCode and MediaInterfaceCode in the CmisHelper.cpp file. The CmisTest.cpp file has also been updated to include the new media interface code in the list of supported applications.

Reviewed By: harshitgulati18

Differential Revision: D86171013

fbshipit-source-id: 3329a0278241c0ea5ec832a52337ce60faac89d9