Add Kernel Version test
Summary: Adds a test to check that the kernel version of the switch is the correct one. Can detect a SEV like S566115 much earlier. TL;DR: somehow certain switches got configured to be provisioned with a different kernel. This caused failures but was only alerted on when some other fboss service failed.

Reviewed By: lohithchittineni

Differential Revision: D85594584

fbshipit-source-id: a290055d96526eec62b64860e9cd4941f82fc3f1