Add packet receiving and validation logic
Summary:
**Context**:
Fabric Link Monitoring provides a way to send packets on specific fabric links to monitor the health of
the links. This needed specific programming on J3 and R3 devices. The sending of packets/cells will
be done from the local CPU of J3 for RDSW/EDSW to FDSW links and the remote R3 will loopback the
packets/cells. For FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link is healthy.
Detailed design of this feature captured here. This stack introduces diffs to support this feature.

Fabric Link Monitoring Manager has the implementation to send/receive fabric link monitoring
packets over fabric ports. On RDSW, we can send these packets on all the fabric link ports without
any restrictions. However, on FDSWs towards SDSW, there is a limit that we can only have 40
outstanding packets per VD, so the number of packets sent will be throttled to ensure there aren't
more outstanding packets. We also are sending packets with payload as 0xA5A5A5A5 and
0x5A5A5A5A to be able to identify any bitflips that can happen in the path though there is not
enough data to conclusively say this can help.

**In this diff**:
Implement packet reception handling with:
- Sequence number validation to detect out-of-order packets
- Port ID verification to ensure packet integrity
- Payload pattern validation alternating pattern
- Automatic detection of dropped packets based on sequence gaps
- Flow control updates to release outstanding packet slots given out of order packets are not
expected

Reviewed By: shri-khare

Differential Revision: D84680031

fbshipit-source-id: fd5d4e8828056d6bb691f7d4f4fbda4e41d1d068