Clean up FBOSS container and add tools for submit checks
The container image was pulling in gcc11 and then later installing
gcc12 and then trying to make users default to it by forcing their
shells to source /opt/rh/gcc-toolset-12/enable. Instead we change things
such that only gcc12 is installed in the image, to avoid any possible
confusion and remove the need for changing PATH.

The container image contained an intermediate layer that had an entire
copy of the FBOSS source tree, just to run getdeps.py install-system-deps
but instead we can only copy the build scripts. We don't typically need
to extract stable hashes there because that's done by the build script
later at build time.

We install clang-format (using the same version used in the GHA) so that
people running vscode inside the container get their code autoformatted
out of the box.