Share PortStatus w/ Transceiver Manager (Port Manager Mode)
Summary:
We previously removed any references to port state stored in TransceiverManager. However, this isn't feasible, because two features require some information about port state.
- Remediation Logic: As it is right now, remediation logic requires knowing which ports are down so it can determine which lanes to reset.
- Because we split transceiver and port states, we can now run into a small edge case in which we  face i2c read failures on a transceiver while the transceiver is inserted and active - TransceiverManager needs some reference to port information to ensure it doesn't trigger fw upgrade while port is still active, because of our upgrade on insert behavior.

There is a long-term plan to remove transceiver remediation from qsfp_service, but we can't do this just yet. So adding this logic in the interim to not break assumptions about functionality.

Acknowledging that this isn't the best design, but once remediation logic is removed, we should be able to just share an `isSafeToRemove` value for each transceiver.

Reviewed By: harshitgulati18

Differential Revision: D85976518

fbshipit-source-id: e4f6cf84e44046a5bd9374e7f5161adfe80eb1d2