Extend EcmpSetupHelper::programRoutes to accept disableTTLDecrement flag
Summary:
Disable TTL decrement on nexthop is 2 steps.
1. Program routes with ECMP and nexthops created. This is the first state update.
2. Update switch state to set disableTTLDecrement=true on the nexthop for every route in the FIBs. This is a 2nd state update.

This worked well until bulk programming is introduced.

Without bulk
- First state update creates ECMP group with nexthop and NHG member objects
- 2nd state update updated the nexthop disableTTLDecrement attribute and deletes/creates ECMP group member object.

At the same time in step 2, because the state changed, the route is also sent down as an update (changeRoute). The expectation here is a new ECMP group is created, route updated and old ECMP deleted. This isn't what is happening.  SaiNextHopGroupManager::incRefOrAddNextHopGroup returns a new SaiNextHopGroupHandle but the actual SAI object just returns the existing NHG object from store since there was no change in key (https://fburl.com/code/5281xsal). This works fine today, although this is unexpected, since we do not alter the state of any ECMP group member other than the single nexthop.
With bulk, this leads to an assertion. Bulk always creates all NHG members. Because we are dealing with existing NHG, the bulk operation asserts since it found an existing NHG member in SaiStore.

One solution to this is to add disableTTLDecrement to nexthop adapterHostKey. That will require fixes to ensure warmboot is supported. It also sets a precedent where we want to update the adapterHostKey for all attributes which are updateable.

Instead, we go with the approach in this stack. Instead of creating an ECMP group with disableTTLDecrement=false, just set it to true within programRoutes. This will avoid the 2nd state update.

This diff enhances the programRoutes API in EcmpSetupHelper to allow configuring the disableTTLDecrement attribute on the nexthops

Reviewed By: msomasundaran

Differential Revision: D86270498

fbshipit-source-id: 6876276bda3b369135209518da4d6a75e3038ab3