Add basic structure of BspPhymanager
Summary:
As titled, this change adds the BspPhymanager header and CPP file.

Most functions, except for getPhyIDInfo and getGlobalXphyID, are currently left unimplemented. Please refer to the upcoming diffs for their implementations.

One notable detail is the MultiPimPlatformSystemContainer* BspSaiPhyManager::getSystemContainer() function. We do not plan to support this, but it must remain because it is a pure virtual function in phymanager.h.

Fortunately, this will not break anything except for CheckPimPresent test, which calls this function. We will address and fix that for ladakh later.

Reviewed By: harshitgulati18

Differential Revision: D85372823

fbshipit-source-id: e67b38b4766ae0fd405204caf2a49cf42bf29725