Initialize scheduler based on switch Ids
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Initialize scheduler to toggle remote neighbors, and then program ECMP routes to measure how long it takes.

Reviewed By: jasmeetbagga

Differential Revision:
D86333711

Privacy Context Container: L1297311

fbshipit-source-id: 74df30a32b02ad7c960551d4e45cfe9763a00956