Properly handle the number of neighbor that's flapping
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Instead of using the interface map with no NDP neighbors, clear out the number of neighbors based on input. That will simulate the number of ports that is flapping on remote device.

Reviewed By: jasmeetbagga

Differential Revision:
D86333718

Privacy Context Container: L1297311

fbshipit-source-id: a4076c6cf8e007b3f350149e96b87cd01b3f4552