Start building HwVoqRouteCompetingRemoteNeighborBenchmark
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Start building benchmark and fix the string formatting.

Reviewed By: jasmeetbagga

Differential Revision:
D86333710

Privacy Context Container: L1297311

fbshipit-source-id: 71eb79345ea461802e9bcbce178ca646760b0d50