Program 16x2K ECMP group while remote neighbor is being updated
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Measure how long it takes to program 16x2K ECMP group while remote neighbor is being updated

Reviewed By: jasmeetbagga

Differential Revision:
D86333714

Privacy Context Container: L1297311

fbshipit-source-id: aa309dcf797421d25a96455d00d03fd3a795d49b