Move benchmark into a benchmark helper
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Move it into a helper file, such that we can configure ECMP group/width, as well as the churn profile (# of ports, # neighbors, churn frequency) for more flexibility during experiment.

Reviewed By: jasmeetbagga

Differential Revision:
D86333712

Privacy Context Container: L1297311

fbshipit-source-id: 57e8e7203c3e887e182d1bf176a2a5f3af1753d8