Fix race condition of scheduler teardown and state clone
Summary:
**Background**

In KCM2 we observed that a rising number of port flaps is causing an uptick number of NSDB state deltas as well as CPHM OOM killed. While NSDB and CPHM issues are being investigated, we need to ensure that agent is able to withstand such high number of neighbor flaps, without depleting all the cycles to update remote neighbor.

To measure the impact of flappy neighbors, create the benchmark to measure the ECMP route programming time when competing with neighbor updates. This would help us to determine if dampening is needed.

**This Diff**
Fixing two issues when trying out multiple inputs for the tests
1. Since we are reapplying two stats back and forth (with and without neighbor), also reset the arp/ndp map to avoid applying an already published state.
2. Race condition observed between scheduling the update vs teardown. Add proper checks when scheduling the update.

Reviewed By: jasmeetbagga

Differential Revision:
D86369125

Privacy Context Container: L1297311

fbshipit-source-id: 3250fa9d96aed00a4f90e728fea0c74c5c80fd50