Release Read lock in creategenerator() before co_yielding to serveStream
Summary:
As titled. If creategenerator() does not release the rlock before co_yielding, this call will be holding on to the lock for longer than necessary.
If serveStream does not invoke gen.next(), its possbile that the rlock will never be released. During S581598, this led to a deadlock in https://fburl.com/code/giwrg9w7 when write call tried to acquire exclusive lock.
Copy over the pubUnit to a local variable and release the rlock as soon as the pipe processing is complete. This will ensure that we will hold on to the lock only when necessary

Differential Revision: D86380732

fbshipit-source-id: 518eef8243b9ff3215438b7dbaf9303bd1e7d200