Add optional profileID argument to getDataPlanePhyChips
Summary:
`getTransceiverIds` as is pulls pin / chip information from static mapping for each port in platform mapping (e.g. https://fburl.com/code/l01d9pq7). Per discussion with Harshit, we don't really need to use this field.

However, in the new platform mapping model where a single port can subsume lanes from the adjacent transceiver, we want to be able to fetch transceiver ids based on what is being used for the current profile.

This diff adds `portProfileID` as an optional argument to `getTransceiverIds` and `getDataPlanePhyChips` in order to facilitate this – similar to other PlatformConfigUtils helpers which have an optional `profileID` field.

If the field is left as nullopt, the function will follow the current behavior.

Reviewed By: harshitgulati18

Differential Revision: D85811806

fbshipit-source-id: 27d3839595cc77d73cfa8c167838e55d4ac8c574