Add rx packet handling for fabric link monitoring
Summary:
**Context**:
Fabric Link Monitoring provides a way to send packets on specific fabric links to monitor the health of
the links. This needed specific programming on J3 and R3 devices. The sending of packets/cells will
be done from the local CPU of J3 for RDSW/EDSW to FDSW links and the remote R3 will loopback the
packets/cells. For FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link is healthy.
Detailed design of this feature captured here. This stack introduces diffs to support this feature.

Fabric Link Monitoring Manager has the implementation to send/receive fabric link monitoring
packets over fabric ports. On RDSW, we can send these packets on all the fabric link ports without
any restrictions. However, on FDSWs towards SDSW, there is a limit that we can only have 40
outstanding packets per VD, so the number of packets sent will be throttled to ensure there aren't
more outstanding packets. We also are sending packets with payload as 0xA5A5A5A5 and
0x5A5A5A5A to be able to identify any bitflips that can happen in the path though there is not
enough data to conclusively say this can help.

**In this diff**:
- Current packet handling always expect an ethernet packet and lookup based on ethertype etc can
help process the packet further. But in fabric link mon, packets are from fabric side and doesnt have
an ethernet header etc., hence needs to be processed differently. Differentiating the processing
based on the fabric port type of packet reception. This needs to be moved to an application type
based processing as requested in CS00012430577 going forward.

Reviewed By: jasmeetbagga

Differential Revision: D85348358

fbshipit-source-id: beb12d38d07c240bc1dcd4fd208d184a6c302398