Add config versions diff tool
Summary:
Have a tool that parses sensor_service configs and compares pm units default sensors with the versioned sensors and output what sensors would need to get modified and for which product versions

Implemented the missing `comparePmUnitSensors` function that was declared in the header but not defined in the cpp file, which was causing linker errors. This function wraps `comparePmSensors` to accept PmUnitSensors objects.

Also fixed a null pointer dereference in `compareVersionedSensors` when the requested pmUnitName doesn't exist in the config - now properly checks if defaultConfig has a value before dereferencing it.

Reviewed By: Scott8440

Differential Revision: D85990915

fbshipit-source-id: b266db67a649e87733cfb035e59281e138b14f04