Move stats to thrift and add thrift handler to get fabric link mon stats for CLI
Summary:
**Context**:
Fabric Link Monitoring provides a way to send packets on specific fabric links to monitor the health of
the links. This needed specific programming on J3 and R3 devices. The sending of packets/cells will
be done from the local CPU of J3 for RDSW/EDSW to FDSW links and the remote R3 will loopback the
packets/cells. For FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link is healthy.
Detailed design of this feature captured here. This stack introduces diffs to support this feature.

Fabric Link Monitoring Manager has the implementation to send/receive fabric link monitoring
packets over fabric ports. On RDSW, we can send these packets on all the fabric link ports without
any restrictions. However, on FDSWs towards SDSW, there is a limit that we can only have 40
outstanding packets per VD, so the number of packets sent will be throttled to ensure there aren't
more outstanding packets. We also are sending packets with payload as 0xA5A5A5A5 and
0x5A5A5A5A to be able to identify any bitflips that can happen in the path though there is not
enough data to conclusively say this can help.

**In this diff**:
- In the next diff, trying to create a CLI to look into the fabric link mon stats, for this, make sure that
stats structure is moved to thrift.
- Add a new thrift handler to get the fabric link monitoring manager stats for use in next diff

Reviewed By: shri-khare

Differential Revision: D85508137

fbshipit-source-id: aedd7ea3ad326ce6247f93d8b0c4324bfed9f7b6