Add fb303 counter for xphyGetPortInfo failed in PhyManager.cpp
Summary:
Adding an fb303 counter to expose xphy getPortInfo calls that fail at this function (https://fburl.com/code/6kirp8u8) - this is typically an indicator that on next warmboot, this device will start to crashloop until a successive coldboot happens to reset communication.

We want to expose these counters as soon as possible, but to tune the thresholds + define proper auto-remediation logic, we need to see how many devices have these issues + how frequently + confirm if this is a valid metric for determining if reboots are necessary to resolve the health of a switch.

Reviewed By: birdsoup

Differential Revision: D86979175

fbshipit-source-id: fe5f899885ecccead47cb440413786e650d67abc