Reduce Unnecessary Port Status Queued Events
Summary:
I noticed an issue with the current port status updating logic where we enqueue PORT_UP / PORT_DOWN events for a state machine when we detect that the new state will be different from the current one. However, we didn't verify that the current state was already PORT_UP or PORT_DOWN. This diff adds that check.

This wouldn't cause any issues as is because none of these events would be applied since their transitions are not supported by the state machine, but it requires more locking and clutters the logs.

Reviewed By: harshitgulati18

Differential Revision: D86492510

fbshipit-source-id: ecc6d9a61acfa48c420f08544df6f7a4291884c1