Fixed `data_corral_service.thrift` link to point to the public FBOSS repository. (#645)
Summary:
The link to `data_corral_service.thrift` in the file `docs/docs/platform/services_support_requirements.md` was pointing to internal private FBOSS repository which was not accessible from outside Meta.  In this PR, changed this link to point to the public FBOSS repository.

Pull Request resolved: https://github.com/facebook/fboss/pull/645

Test Plan: N/A

Reviewed By: shri-khare

Differential Revision: D87182477

Pulled By: harshitgulati18

fbshipit-source-id: 938a9e297964397d8af9ab8116fd37ac4a2bfe76