Add clang build support
FBOSS OSS can now optionally be built with clang by passing a new
`--use-clang` flag to the script script.  When that flag is passed
the FBOSS container will default using clang however the command
`use-gcc` can be used to switch back to gcc12 if needed, and the
command `use-clang` can then be used to switch back to clang. The
same container can be used to build FBOSS with both compilers by
switching back and forth between the two seamlessly.

Note that the same gcc-built SAI/SDK can be used when building FBOSS
with clang. It's not necessary to rebuild the SAI/SDK with clang.

- Add `--use-clang` flag to `docker-build.py` to build with LLVM toolchain,
  in which case we put a full blown clang install in the Docker image
  and configure it as the default compiler.
- Auto-detect clang in `run-getdeps.py` and set compiler flags needed as clang
  tends to be stricter than gcc. Re-use compiler flags from `CMakeLists.txt`
  to avoid duplication.
- Make sure we don't install `binutils` as the entire toolchain is either
  provided by gcc12 or by clang, and binutils creates some paths such as
  `/usr/bin/ar` that aren't symlinks and thus can't be swapped out by using
  `update-alternatives`. Doing this requires building `libiberty` locally,
  which requires patching the existing Meta patch (insert yo dawg meme here)
  so that it works with clang.
- Prepend clang flags to user environment variables for override capability
- Link using clang's linker, `lld`, which is way faster (more than 6x!).
- Fix `qsfp_hw_test` duplicate `main()` by using `fboss_init` instead of
  `hw_test_main` as gcc was accepting to link this incorrect program where
  the binary has two `main()` functions.
- Disable range-v3 tests and perf benchmarks to speed up builds and work
  around some clang specific build failures that don't matter to FBOSS.
- Fix insufficient forward declaration and unused variable warnings.