Add a new feature for buffer pool size not including headroom size
Summary:
It is left for implementations to decide if headroom pool size is part of the ingress buffer pool size or
not. We have some implementations like Chenab where the ingress pool size is same as the shared
buffer size without including the headroom pool size, which is different from rest of the implementations.
Hence, adding a new feature to guard this.

Reviewed By: shri-khare

Differential Revision: D87080592

fbshipit-source-id: ad00b721cbba413c9d7e58a92e43d80f4f6d5070