Avoid headroom size to be part of the pool size for Chenab
Summary:
It is left for implementations to decide if headroom pool size is part of the ingress buffer pool size or
not. We have some implementations like Chenab where the ingress pool size is same as the shared
buffer size without including the headroom pool size, which is different from rest of the implementations.
Using the new feature to avoid including headroom size in the overall pool size.

Reviewed By: shri-khare

Differential Revision: D87042331

fbshipit-source-id: 69029cd13f0a610bee3af0bf88cd7bf871ecd241