Align fw_util spi_layout for BIOS verify and read commands with upgrade command (#576)
Summary:
This PR addresses a bug where the `spi_layout` was missing from the BIOS `verify` and `read` command configurations in the fw_util config files. This caused the layout for these commands to not match the one used by the upgrade command. This issue was introduced during a refactoring a few months ago, when the `preUpgradeCmd` was removed.

### Testing
Verified BIOS read and verify commands on darwin, darwin48v, meru800bfa and meru800bia systems.

Pull Request resolved: https://github.com/facebook/fboss/pull/576

Reviewed By: somasun

Differential Revision: D84928717

Pulled By: joancaneus

fbshipit-source-id: c0de07c67657e4af5c6a72fd9e4e155ff790a961