is200gFr4InLowPowerMode vendor specific and OSS implementation
Summary:
The diff handles the special handling for the finisar 200G_FR4 optics. The special handling is required for finisar optics as we have observed on the fleet on minipack2 with finisar operating in 100G,  that Tx-LOL alarm is consistent on the host side. We are coming with the below seqeuence based on the multiple iteration of various programming sequence. Finisar is actively debugging the issue for potentail f.w fix.

-> The module is kept at low power mode in the ready state machine state
-> Utility functions has been introduced for better code structuring
-> 200G_FR4 optics is moved to high power mode only after completion of AppSel programming
-> Poll is introduced for verifying the module in ready state.

Reviewed By: harshitgulati18

Differential Revision: D87202067

fbshipit-source-id: 1d7219f7bbac7f2ecd6fbaada6b61468b2fe5f1c