Fix `show transceiver` for multi-port modules
Summary:
shiva-menta encountered another bug with the recent changes I made to support bypass modules in the fboss2 CLI: P2037702538. In this paste we see 6 interfaces listed as bypass modules, despite being on an actual optic.

When a module is in 2x800G mode, the TransceiverInfo still lists all interfaces that are possible if it were to run in 8x200G mode. Since the agent only has 2 ports configured instead of 8, the current logic assumes that the other 6 ports are for bypass modules.

This diff modifies the logic a bit so that we will only consider a transceiver to be a bypass module if NONE of the interfaces on the transceiver have an associated agent port. For non-bypass-modules we will skip any interfaces that don't have an associated agent port.

Differential Revision: D86991187

fbshipit-source-id: 1f9be2bcdb6de07cc9ae7d3f8088ccac40801bdb