Add FibsInfoMap to switch state.
Summary: This diff stack will contain changes to move the FIB in the current switch state to a newer design. The design will have a new map FibsInfoMap, which is a map<SwitchIdList, FibInfoFields>. FibsInfoFields is a struct which contains the existing map<i16, FibContainerFields>. This is done to accommodate more FIB related maps in the future in this new struct. As part of ECMP nexthop migration, we can store the NextHop ID Maps in the FibInfoFields struct. I will add that diff later here.

Reviewed By: zechengh09

Differential Revision:
D86916049

Privacy Context Container: L1297311

fbshipit-source-id: 42e12e856cb5a9ed478cf2bc651beae2d6b547e8