Add support to pass in PG stats that are supported to reduce fb303 keys
Summary:
We have both congestionDiscards and congestionDiscardSeen counters and need a way to ensure
that we don't expose both the counters if PFC is enabled for a port. As of now, these are mutually
exclusive, but passing both the capabilities on a per port basis to ensure we limit counter export
based on if  congestionDiscards is supported or if congestionDiscardSeen is supported.

Differential Revision: D87473612

fbshipit-source-id: fcb36938215f1aa00e2f16d752b6fbbaaaa9cb33